add: support postgresql dialect

+++ tablemap.go
@@ -27,6 +27,7 @@ type tableMapOption struct {
 	uniqueWithName     bool
 	foreignKeyWithName bool
 	outerForeignKey    bool
+	withoutDropTable   bool
 }
 
 type tableMapOptionFunc func(*tableMapOption)
@@ -292,11 +293,13 @@ func retrieveIndexForeignKeyOptionByExpr(exprs []ast.Expr) []index.ForeignKeyOpt
 	return options
 }
 
-func (tm *TableMap) WriteDDL(w io.Writer, dialect Dialect) error {
+func (tm *TableMap) WriteDDL(w io.Writer, dialect Dialect, opts *tableMapOption) error {
 	tableName := dialect.QuoteField(strings.TrimSpace(tm.Name))
 
-	if _, err := io.WriteString(w, "DROP TABLE IF EXISTS "+tableName+";\n\n"); err != nil {
-		return fmt.Errorf("failed to write drop table string: %w", err)
+	if !opts.withoutDropTable {
+		if _, err := io.WriteString(w, "DROP TABLE IF EXISTS "+tableName+";\n\n"); err != nil {
+			return fmt.Errorf("failed to write drop table string: %w", err)
+		}
 	}
 	if _, err := io.WriteString(w, "CREATE TABLE "+tableName+" (\n"); err != nil {
 		return fmt.Errorf("failed to write create table string: %w", err)
